jets: rename `crypto` to `lock`
